# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: CraftLand Hub API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /submissions:
    get:
      operationId: listSubmissions
      tags: [submissions]
      summary: List submissions (admin only)
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, approved, rejected]
      responses:
        "200":
          description: List of submissions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Submission"
    post:
      operationId: createSubmission
      tags: [submissions]
      summary: Submit a new map
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSubmissionBody"
      responses:
        "201":
          description: Submission created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Submission"
  /submissions/{id}/approve:
    post:
      operationId: approveSubmission
      tags: [submissions]
      summary: Approve a submission (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Submission approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Submission"
  /submissions/{id}/reject:
    post:
      operationId: rejectSubmission
      tags: [submissions]
      summary: Reject a submission (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RejectSubmissionBody"
      responses:
        "200":
          description: Submission rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Submission"

  # Users
...
  /reports:
    get:
      operationId: listReports
      tags: [reports]
      summary: List all reports (admin only)
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, resolved, dismissed]
      responses:
        "200":
          description: Reports list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Report"
    post:
      operationId: createReport
      tags: [reports]
      summary: Report a map
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateReportBody"
      responses:
        "201":
          description: Report created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Report"
  /reports/{id}/resolve:
    post:
      operationId: resolveReport
      tags: [reports]
      summary: Resolve a report (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Report resolved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Report"
  /reports/{id}/dismiss:
    post:
      operationId: dismissReport
      tags: [reports]
      summary: Dismiss a report (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Report dismissed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Report"

  # Analytics
...
components:
  schemas:
...
    Submission:
      type: object
      properties:
        id:
          type: integer
        creatorName:
          type: string
        region:
          type: string
        mapCode:
          type: string
        mapLink:
          type: ["string", "null"]
        image:
          type: ["string", "null"]
        mapName:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, approved, rejected]
        rejectionReason:
          type: ["string", "null"]
        submittedByUserId:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - creatorName
        - region
        - mapCode
        - status
        - createdAt
    CreateSubmissionBody:
      type: object
      properties:
        creatorName:
          type: string
        region:
          type: string
        mapCode:
          type: string
        mapLink:
          type: ["string", "null"]
        image:
          type: ["string", "null"]
        mapName:
          type: ["string", "null"]
      required:
        - creatorName
        - region
        - mapCode
    RejectSubmissionBody:
      type: object
      properties:
        reason:
          type: ["string", "null"]
...
    Report:
      type: object
      properties:
        id:
          type: integer
        mapId:
          type: integer
        mapName:
          type: ["string", "null"]
        userId:
          type: ["string", "null"]
        reason:
          type: string
          enum: [spam, invalid_code, inappropriate_content, other]
        description:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, resolved, dismissed]
        createdAt:
          type: string
      required:
        - id
        - mapId
        - reason
        - status
        - createdAt
    CreateReportBody:
      type: object
      properties:
        mapId:
          type: integer
        reason:
          type: string
          enum: [spam, invalid_code, inappropriate_content, other]
        description:
          type: ["string", "null"]
      required:
        - mapId
        - reason
...